requirements.txt
temporalio
fastapi
uvicorn
asyncpg
sqlalchemy
alembic
pytest
python-dotenv
aiomysql>=1.1.0
